Solo falta arreglar un controller

+++ desafio2-dwf/src/main/java/sv/edu/udb/repository/CargoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import sv.edu.udb.domain.Cargo;
 
 @Repository
-public interface CargoRepository extends JpaRepository<Cargo, Integer> {
+public interface CargoRepository extends JpaRepository<Cargo, Long> {
 }
